move der_local_values to it's own file
