+++ netket/operator/__init__.py
@@ -1,5 +1,8 @@
 from .local_values import (
     local_values,
+)
+
+from .der_local_values import (
     der_local_values,
 )
